requirements.txt
streamlit
openai
sentence-transformers
numpy
faiss-cpu
beautifulsoup4